Refactoring KDE class out of AnalysisGraph.hpp
